backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.4
langchain-community>=0.0.5
ddgs>=4.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-docx>=1.1.0
